+++ web-wallet/src/routes/(app)/dashboard/+page.svelte
@@ -4,9 +4,9 @@
   import { onDestroy } from "svelte";
   import { fade } from "svelte/transition";
   import { filterWith, find, hasKeyValue, last } from "lamb";
-  import { mdiDatabaseOutline, mdiSwapVertical } from "@mdi/js";
+  import { mdiContain, mdiDatabaseOutline, mdiSwapVertical } from "@mdi/js";
 
-  import { Tabs } from "$lib/dusk/components";
+  import { Icon, Tabs } from "$lib/dusk/components";
   import { StakeContract, TransferContract } from "$lib/containers";
   import { AddressPicker, Balance, Transactions } from "$lib/components";
   import { operationsStore, settingsStore, walletStore } from "$lib/stores";
@@ -52,6 +52,7 @@
     id,
     label,
   }));
+  const hasNoEnabledContracts = enabledContracts.length === 0;
 
   let selectedTab = tabItems[0]?.id ?? "";
 
@@ -76,6 +77,21 @@
     tokenCurrency="DUSK"
     tokens={balance.value}
   />
+
+  {#if hasNoEnabledContracts}
+    <div class="no-contracts-pane">
+      <Icon path={mdiContain} size="large" />
+      <h3>No Contracts Enabled</h3>
+      <p>
+        It appears that no contracts are currently enabled. To access the full
+        range of functionalities, enabling contracts is essential.
+      </p>
+      {#if import.meta.env.MODE === "development"}
+        <h4>For Developers:</h4>
+        <p>No contracts are currently enabled.</p>
+      {/if}
+    </div>
+  {/if}
 
   {#if selectedContract}
     <article class="tabs">
@@ -145,6 +161,32 @@
       flex-direction: column;
       padding: 1rem 1.375rem;
       gap: var(--default-gap);
+    }
+  }
+
+  .no-contracts-pane {
+    display: flex;
+    flex-direction: column;
+    background-color: var(--surface-color);
+    padding: 1rem 1.375rem;
+    border-radius: var(--control-border-radius-size);
+
+    & h3 {
+      text-align: center;
+      margin-bottom: 1em;
+    }
+
+    & p:not(:last-child) {
+      margin-bottom: 1em;
+    }
+
+    h4 {
+      margin-bottom: 0.5em;
+    }
+
+    :global(.dusk-icon) {
+      align-self: center;
+      margin-bottom: 0.5rem;
     }
   }
 </style>